requirements.txt
aiogram==3.0.0
tortoise-orm==0.19.3
pydantic==2.3.0
pydantic-settings==2.0.2
